backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
spotipy==2.23.0
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
langdetect==1.0.9
